// cms/src/data/types.ts
export interface Config {
  title: string;
  description: string;
  idField: string,
  columns: string[];
  editRoute?: (rowData: any) => Record<string, any>;
}


export type CoordinatesFrequency = {
  coordinates: [number, number],
  count: number,
  images: string[]
}
export interface ErrorResponseData {
  message?: string; // Optional, because not all error responses may contain a message
}

export type ComponentType = 'biosample' | 'organism' | 'assembly' | 'localSample' | 'experiment' | 'annotation';

export type DataModels = 'biosamples' | 'experiments' | 'organisms' | 'annotations' | 'assemblies' | 'local_samples'